Update UX for general facility detail fields
